splat-to-mesh/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
requests==2.31.0
google-cloud-storage==2.14.0
python-multipart==0.0.6

